+++ src-docs/src/views/i18n/i18n_basic.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import {
+  EuiI18n,
+} from '../../../../src/components';
+
+export default () => {
+  return (
+    <p>
+      <EuiI18n
+        token="i18n.basicexample"
+        default="This is the English copy that would be replaced by a translation defined by the i18n.basicexample token."
+      />
+    </p>
+  );
+};