+++ src/components/index.js
@@ -61,6 +61,11 @@ export {
 } from './combo_box';
 
 export {
+  EuiContext,
+  EuiI18nConsumer
+} from './context';
+
+export {
   EuiContextMenu,
   EuiContextMenuPanel,
   EuiContextMenuItem,
@@ -180,6 +185,11 @@ export {
 export {
   EuiImage,
 } from './image';
+
+export {
+  EuiI18n,
+  EuiI18nNumber,
+} from './i18n';
 
 export {
   EuiLoadingKibana,